refactor(reference-data): split registry postgres_registry.go by domain
Split the 808-line registry into focused files:
- registry_instrument.go: CRUD operations (Get, List, Create, Update, Activate, Deprecate)
- registry_validation.go: CEL validation, compilation, caching
- postgres_registry.go: struct, constructor, transaction helpers, scan helpers
